pyproject.toml
[project]
name = "financial-intelligence-agent"
version = "1.0.0"
description = "Stateful AI agent for Apple 10-K financial analysis"
requires-python = ">=3.11"

dependencies = [
    # LangGraph — stateful agent orchestration
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",

    # Google Gemini — LLM + embeddings
    "google-genai>=0.8.0",
    # Local embeddings — ONNX-based, no PyTorch needed
    "fastembed>=0.4.0",

    # Vector store
    "chromadb>=0.5.0",

    # Web search
    "tavily-python>=0.3.0",

    # HTML parsing (upstream pipeline)
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",

    # API server
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "typing-extensions>=4.11.0",
]

[project.optional-dependencies]
redis = ["redis>=5.0.0"]    # uv add --optional redis

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["agent", "ingest", "server", "eval"]

[tool.ruff]
line-length = 100
target-version = "py311"